feat: renamed edc service to connector service
