+++ INSTALL.md
@@ -36,9 +36,9 @@ pip install --upgrade tractusx-sdk
 Here's a quick example of how to use the SDK:
 
 ```python
-from tractusx_sdk.dataspace.services.connector.v0_9_0.edc_service import EdcService
+from tractusx_sdk.dataspace.services.connector.v0_9_0.edc_service import ConnectorService
 
-edc_service = EdcService(
+edc_service = ConnectorService(
     base_url="https://control.plane.url", 
     dma_path="management",
     headers={
