+++ src/tractusx_sdk/dataspace/services/__init__.py
@@ -24,6 +24,6 @@
 __author__ = 'Eclipse Tractus-X Contributors'
 __license__ = "Apache License, Version 2.0"
 
-from .connector.base_edc_service import BaseEdcService
+from .connector.base_connector_services import BaseConnectorService
 
-__all__ = ['BaseEdcService']
+__all__ = ['BaseConnectorService']
